fix(toolbar): extract toolbar content logic into a custom hook and it reuse in short tollbar state
