+++ src/features/toolbar/components/ToolbarContent/ToolbarContent.tsx
@@ -1,35 +1,12 @@
-import { Fragment, useMemo } from 'react';
-import { useAtom } from '@reatom/react-v2';
+import { Fragment } from 'react';
 import { ChevronDown16 } from '@konturio/default-icons';
-import { toolbar } from '~core/toolbar';
+import { useToolbarContent } from '~features/toolbar/hooks/use-toolbar-content';
 import { ToolbarControl } from '../ToolbarControl/ToolbarControl';
 import { ToolbarButton } from '../ToolbarButton/ToolbarButton';
 import s from './ToolbarContent.module.css';
-import type { ControlState, ToolbarControlSettings } from '~core/toolbar/types';
-import type { PrimitiveAtom } from '@reatom/core-v2/primitives';
-
-type ToolbarControlButton = {
-  id: string;
-  settings: ToolbarControlSettings;
-  stateAtom: PrimitiveAtom<ControlState>;
-};
 
 export const ToolbarContent = () => {
-  const [controls] = useAtom(toolbar.controls);
-
-  const toolbarContent = useMemo(() => {
-    return toolbar.toolbarSettings.sections
-      .map((section) => {
-        const buttons = section.controls.reduce<ToolbarControlButton[]>((acc, id) => {
-          const settings = controls.get(id);
-          const stateAtom = toolbar.getControlState(id);
-          if (settings && stateAtom) acc.push({ id, settings, stateAtom });
-          return acc;
-        }, []);
-        return { sectionName: section.name, buttons };
-      })
-      .filter(({ buttons }) => buttons.length > 0);
-  }, [controls]);
+  const toolbarContent = useToolbarContent();
 
   return (
     <div className={s.toolbarContent}>
